feat: add the autocompletion feature to StringableInputComponent
